+++ src/reanimated2/commonTypes.ts
@@ -16,6 +16,7 @@ import {
   TextStyle,
 } from 'react-native';
 import { Animation, AnimationObject } from './animation/commonTypes';
+import { Context } from './hook/commonTypes';
 
 export type TransformProperty =
   | PerpectiveTransform
@@ -45,4 +46,20 @@ export interface AnimatedStyle
 }
 export interface SharedValue<T> {
   value: T;
+}
+
+export interface WorkletFunction {
+  _closure?: Context;
+  __workletHash?: number;
+  __optimalization?: number;
+  __worklet?: boolean;
+}
+
+export interface BasicWorkletFunction<T> extends WorkletFunction {
+  (): T;
+}
+
+export interface ComplexWorkletFunction<A extends any[], R>
+  extends WorkletFunction {
+  (...args: A): R;
 }
