+++ src/reanimated2/hook/commonTypes.ts
@@ -1,16 +1,6 @@
 export type DependencyList = Array<unknown> | undefined;
 
 export type Context = Record<string, unknown>;
-export interface WorkletFunction {
-  _closure?: Context;
-  __workletHash?: number;
-  __optimalization?: number;
-  __worklet?: boolean;
-}
-
-export interface BasicWorkletFunction<T> extends WorkletFunction {
-  (): T;
-}
 
 export interface ContextWithDependencies<TContext extends Context> {
   context: TContext;
