+++ src/reanimated2/hook/useDerivedValue.ts
@@ -1,8 +1,8 @@
 import { useEffect, useRef } from 'react';
 import { initialUpdaterRun } from '../animation';
-import { SharedValue } from '../commonTypes';
+import { BasicWorkletFunction, SharedValue } from '../commonTypes';
 import { makeMutable, startMapper, stopMapper } from '../core';
-import { BasicWorkletFunction, DependencyList } from './commonTypes';
+import { DependencyList } from './commonTypes';
 
 export type DerivedValue<T> = Readonly<SharedValue<T>>;
